http://issues.carrot2.org/browse/HPPC-27
The benchmark was designed so that access to a volatile variable was within the loop. This caused mem bariers inside the jitted code and (my suspicion based on assembly dumps) turned off certain optimisations that otherwise are taking place. I changed the code to use thread-local processing with assertions at the end, seems to work fine and shows that all iterators have pretty much the same speed. This result may again be skewed because inlined bytecode size for get() differs a lot from the bytecode size of IntCursor iterator... but for tight loops all of them should work identically.

git-svn-id: https://carrot2.svn.sourceforge.net/svnroot/carrot2/labs/hppc/trunk/hppc-core@4263 7ff1d41c-760d-0410-a7ff-a3a56f310b35